chore: electron based console
