+++ packages/console/dev-app-update.yml
@@ -1,0 +1,3 @@
+provider: generic
+url: https://example.com/auto-updates
+updaterCacheDirName: console-updater
