+++ packages/console/package.json
@@ -1,45 +1,57 @@
 {
-  "name": "@hot-updater/console",
-  "type": "module",
-  "private": true,
+  "name": "@hot-update/console",
+  "description": "React Native OTA solution for self-hosted",
+  "license": "MIT",
+  "repository": "https://github.com/gronxb/hot-updater",
+  "author": "gronxb <gron1gh1@gmail.com> (https://github.com/gronxb)",
+  "bugs": {
+    "url": "https://github.com/gronxb/hot-updater/issues"
+  },
+  "homepage": "https://github.com/gronxb/hot-updater#readme",
   "version": "0.0.1",
-  "main": "./dist/main.mjs",
+  "main": "./out/main/index.js",
   "scripts": {
-    "build": "rsbuild build --mode production",
-    "check": "biome check --write",
-    "dev": "rsbuild dev --open",
-    "format": "biome format --write",
-    "preview": "rsbuild preview",
-    "dev:gui": "concurrently \"cross-env NODE_ENV=development rsbuild dev\" \"wait-on tcp:3000 && cross-env NODE_ENV=development electron main.js\"",
-    "preview:gui": "cross-env NODE_ENV=production electron main.js"
+    "typecheck:node": "tsc --noEmit -p tsconfig.node.json --composite false",
+    "typecheck:web": "tsc --noEmit -p tsconfig.web.json --composite false",
+    "typecheck": "npm run typecheck:node && npm run typecheck:web",
+    "start": "electron-vite preview",
+    "dev": "electron-vite dev",
+    "build": "npm run typecheck && electron-vite build",
+    "postinstall": "electron-builder install-app-deps",
+    "build:unpack": "npm run build && electron-builder --dir",
+    "build:win": "npm run build && electron-builder --win",
+    "build:mac": "electron-vite build && electron-builder --mac",
+    "build:linux": "electron-vite build && electron-builder --linux"
   },
   "dependencies": {
-    "@hot-updater/core": "workspace:^",
-    "@hot-updater/plugin-core": "workspace:^"
+    "@electron-toolkit/preload": "^3.0.1",
+    "@electron-toolkit/utils": "^3.0.0",
+    "electron-updater": "^6.1.7"
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.8.3",
+    "@electron-toolkit/eslint-config-prettier": "^2.0.0",
+    "@electron-toolkit/eslint-config-ts": "^2.0.0",
+    "@electron-toolkit/tsconfig": "^1.0.1",
     "@kobalte/core": "^0.13.6",
-    "@rsbuild/core": "^1.0.9",
-    "@rsbuild/plugin-babel": "^1.0.1",
-    "@rsbuild/plugin-solid": "^1.0.2",
     "@solidjs/meta": "^0.29.4",
     "@solidjs/router": "^0.14.7",
     "@tanstack/solid-table": "^8.20.5",
+    "@types/node": "^20.14.8",
     "autoprefixer": "^10.4.19",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "concurrently": "^9.0.1",
-    "cross-env": "^7.0.3",
-    "electron": "^32.1.2",
+    "electron": "^31.0.2",
+    "electron-builder": "^24.13.3",
+    "electron-vite": "^2.3.0",
     "lucide-solid": "^0.446.0",
     "postcss": "^8.4.38",
-    "solid-js": "^1.8.22",
+    "prettier": "^3.3.2",
+    "solid-js": "^1.8.17",
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "^3.4.4",
     "tailwindcss-animate": "^1.0.7",
-    "tsup": "^8.3.0",
     "typescript": "^5.5.2",
-    "wait-on": "^8.0.1"
+    "vite": "^5.4.8",
+    "vite-plugin-solid": "^2.10.2"
   }
 }
