--- packages/console/preload.js
@@ -1,6 +1,0 @@
-const { contextBridge, ipcRenderer } = require("electron");
-
-contextBridge.exposeInMainWorld("app", {
-  getAppVersion: () => "Hello",
-  getUpdateJson: () => ipcRenderer.invoke("getUpdateJson"),
-});
