--- packages/console/src/env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@rsbuild/core/types" />
