+++ packages/console/src/renderer/src/main.tsx
@@ -1,0 +1,6 @@
+import "./app.css";
+
+import { render } from "solid-js/web";
+import App from "./App";
+
+render(() => <App />, document.getElementById("root") as HTMLElement);
