+++ packages/console/src/renderer/src/routes/empty-config.tsx
